Add Tags component
This component takes in an array of strings (the tags) and optionally a link, name, and max.
The component will then display the tags as little orange buttons. If there are more tags than max it will cut off any tags of index greater than that and display a hyperlink with text 'name' and href 'link'. This can be used to direct the user to a full list of tags. By default max is set to the length of tags and will display all of them.