feature(app) Change the Results and TagResults components to take explicit booleans on whether to show the 'show more' button, pagination, and a link at the top instead of a count.
Also allow user to pass in hide as a boolean that determines whether to show the results or not. This is used for the different search pages, as well as hiding things when they have no results.

As well as that allow a containerClass to be passed in that will add a class to the div that holds the results.