bugfix(api) Fix instances where data returned from the search api was empty or producing an error
product_data_hash was producing empty data_sizes even when there was data present.
When the search string was empty, Fletcher::Search.related_tags was returning an error due to @products_related_tags being empty since where it was created (.products) was returning before it was created. Now we just return a blank array if the search_String is empty.