fix(app/tests) SharedConnection only used for feature tests.
SharedConnection is only used for feature tests now which should
not force all threads to use the same connection for unit tests.

This module also means that we should be able to use transactional
fixtures for feature tests which should speed stuff up.

Left the option to set false in db_cleaner_features commented out
in case we have conflicts later.