feature(app) Move storage of Collections out of the UserStore and ProductStore and into CollectionsStore. Add Collection editing/sharing to product and collections page.
Previously the collections for a user or product were being stored on the User and Product stores. This made it awkward to edit them from anywhere but the user page. Now they have been moved into the collections store along with the collection is currently being edited. The collections modal now gets data from the store, rather than being passed it by the user or product pages.