feature(api) Improve speed and reduce total SQL queries made durin API requests
Various pieces of information could be cached to reduce the total queries being made on each API request. Need to keep these low and keep lookin in the future for places to reduce the number of queries, and API calls, to keep the app running quickly. Also removes various things from API responses that just didn't need to be sent along.