fix(server) Improve webpack config for dev
On some devices, you can have a hard time getting Webpack to pick up
your changes at a reasonable rate. These are the settings that have
most reliably allowed all of my machines to update.

In addition, I have disabled the hot-swap functionality for now, as
it is currently just a waste of resources if we aren't using it.

Side effects:

- Less console spam

- More productivity

- A little more resource usage