fix(ManageCollab) Make sure remove button works
Upon removing a user from the list, the desktop version of the Manage
Collaborator Collection Modal did not display the change, despite ac-
tually removing the user. This change alters the user list of the modal
to be based on the unsaved collection user list, which solves it.

I did not observe any side effects from this change, but do let me know
if other components are affected.