Add in Single Table Inheritence system for Columns
This means we can have multiple types of column on the same relation/table.
This way when you call generator.columns you will get a collection of all the columns, some are one type some are the other. The system is then polymorphic so that when you call column.process on any of those columns it will call that column subclass' process and pick functions to create that type of column. Pretty cool. Proper OOP inheritence in rails models.