Make it easier to create columns
This before_save made it very difficult to create columns as you had to create them and then immediately have the upcoming join tables availabel in the database. This was not possible.

By moving this to after save you’re able to create the column and then link it to its parent generator.