Add system for importing generators
The JSON hashes for these games comes from commit ef0860a5, this is basically a reversion of that with a cleanup.

Looking back on this, it’d be better to just use these json blobs as the data store rather than the database. I may revert back to using that since it’d speed up response times and the user generated scenarios seems unlikely to be produced.